requirements.txt
# --- Core Data & Dashboard Libraries ---
altair>=5.5.0
numpy>=2.2.0
pandas>=2.3.0
pyarrow>=21.0.0
SQLAlchemy>=2.0.0
streamlit>=1.51.0

# --- Database Drivers ---
psycopg2-binary>=2.9.11
python-dotenv>=1.2.1

# --- Jupyter / IPython Support ---
ipykernel>=7.1.0
ipython>=8.37.0
jupyter_client>=8.6.3
jupyter_core>=5.9.1
matplotlib-inline>=0.2.1

# --- Supporting Libraries & Dependencies ---
asttokens>=3.0.0
attrs>=25.4.0
blinker>=1.9.0
cachetools>=6.2.1
certifi>=2025.10.5
charset-normalizer>=3.4.4
click>=8.3.0
comm>=0.2.3
debugpy>=1.8.17
decorator>=5.2.1
exceptiongroup>=1.3.0
executing>=2.2.1
gitdb>=4.0.12
GitPython>=3.1.45
idna>=3.11
jedi>=0.19.2
Jinja2>=3.1.6
jsonschema>=4.25.1
jsonschema-specifications>=2025.9.1
MarkupSafe>=3.0.3
narwhals>=2.10.1
nest-asyncio>=1.6.0
packaging>=25.0
parso>=0.8.5
pillow>=12.0.0
platformdirs>=4.5.0
prompt_toolkit>=3.0.52
protobuf>=6.33.0
psutil>=7.1.3
pure_eval>=0.2.3
pydeck>=0.9.1
Pygments>=2.19.2
python-dateutil>=2.9.0
pytz>=2025.2
pyzmq>=27.1.0
referencing>=0.37.0
requests>=2.32.5
rpds-py>=0.28.0
six>=1.17.0
smmap>=5.0.2
stack-data>=0.6.3
tenacity>=9.1.2
toml>=0.10.2
tornado>=6.5.2
traitlets>=5.14.3
typing_extensions>=4.15.0
tzdata>=2025.2
urllib3>=2.5.0
wcwidth>=0.2.14